Time: 0 ms (100.00%) | Memory: 44.9 MB (100.00%) - LeetSync

+++ 3787-maximum-students-on-a-single-bench/maximum-students-on-a-single-bench.cpp
@@ -3,7 +3,7 @@ class Solution {
 public:
     int maxStudentsOnBench(vector<vector<int>>& st) {
         if(st.size()==0) return 0;
-        int hm[sz][sz] = {},k[sz] = {},ans = INT_MIN;
+        short hm[sz][sz] = {},k[sz] = {},ans = INT_MIN;
         for(auto& r:st){
             int s = r[0], b =r[1];
             if(hm[s][b])continue;
